Make disassembler work on whole program instead of functions
Previously, the disassembler would work on a function at a time,
using Function.base_address to determine the initial address. Modified
it to disassemble the finished instruction array, which is more true
to what was actually generated. To annotate the function start
addresses, I compare the program counter against the function list.